Extract common receive to separate fragment
