+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/receive/viewmodels/ReceiveSharedViewModel.kt
@@ -3,13 +3,11 @@ package io.horizontalsystems.bankwallet.modules.receive.viewmodels
 import androidx.lifecycle.ViewModel
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.entities.Account
-import io.horizontalsystems.bankwallet.entities.Wallet
 import io.horizontalsystems.bankwallet.modules.receive.ui.UsedAddressesParams
 import io.horizontalsystems.marketkit.models.FullCoin
 
 class ReceiveSharedViewModel : ViewModel() {
 
-    var wallet: Wallet? = null
     var coinUid: String? = null
     var usedAddressesParams: UsedAddressesParams? = null
